Turn on live data settings when changing setting
